backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0
google-generativeai==0.3.2
sentence-transformers==2.2.2

# Research enhancements
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.2
setfit==1.0.3
datasets==2.16.1
scikit-learn==1.3.2
seqeval==1.2.2
optimum==1.16.1
onnxruntime==1.16.3
